Move config to method get() and optimize
Simplify http configurations by putting it in the method `get()`
instead of having an external file for it.

In method `get()` in shortcuts.py, we removed the `timeout` and
`max_request` arguments in order to have same configurations which is
declared inside the method. We add a `timeout` argument in the
request.get library to handle the request and response wait time.